+++ packages/electron-esbuild/src/config/configurators/base.configurator.ts
@@ -6,14 +6,14 @@
 
 import type { EnvConfig } from '../config.js'
 import { Target, TypeConfig } from '../enums.js'
-import { ConfigMapping } from '../types.js'
+import { ConfigMapping, ExtractArray } from '../types.js'
 
 export interface Configurator<P extends TypeConfig> {
   readonly type: TypeConfig
   readonly config: EnvConfig
 
   toBuilderConfig(
-    partial: Partial<ConfigMapping[P]>,
+    partial: Partial<ExtractArray<ConfigMapping[P]>>,
     config: ConfigMapping[P],
     target: Target,
   ): ConfigMapping[P]